refactor: API文档抽离为新目录和路由
