+++ src/dashboard-front/src/views/apiDocs/components/doc-detail-main-content.vue
@@ -67,7 +67,7 @@ import {
   INavItem,
   IResource,
   TabType,
-} from '@/views/apigwDocs/types';
+} from '@/views/apiDocs/types';
 import {
   computed,
   inject,
