+++ src/dashboard-front/src/views/apiDocs/types.d.ts
